+++ src/common/i18n/fr.js
@@ -21,7 +21,9 @@ Roles.i18n = {
             panels: {
                 add_button: 'Nouveau périmètre',
                 add_title: 'Ajoute un nouveau périmètre auqle vous pourrez attacher des rôles spécifiques',
-                edit_no_scope: 'Aucun rôle n\'est encore attribué sur aucun périmètre.<br />Commencez par ajouter un nouvau périmètre.',
+                edit_no_scope: 'Aucun rôle n\'est encore attribué sur aucun périmètre.',
+                edit_no_scope_add: 'Commencez par ajouter un nouvau périmètre.',
+                edit_no_scope_none: 'Aucun périmètre n\'est malheureusement actuellement disponible.',
                 remove_scope_confirm: 'Vous êtes sur le point de supprimer le périmètre "%s", ainsi que tous les rôles qui lui sont attachés.<br />Etes-vous sûr ?',
                 remove_scope_title: 'Suppression de périmètre',
                 remove_title: 'Supprime tous les rôles de ce périmètre',