+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define Roles.suggestedPermissions() new function
     - Define Roles.compareLevels() new function
     - Reinit scopes before reloading them
+    - Have distinct behavior when no scoped role depending of whether there are available scopes
 
 ### 1.5.0